requirements.txt
-i https://pypi.org/simple
astroid==2.3.3
attrdict==2.0.1
attrs==19.3.0
babel==2.7.0
base58==1.0.3
certifi==2019.11.28
chardet==3.0.4
cytoolz==0.10.1
django-countries==5.5
django-monthfield==0.1.2
django-phonenumber-field==3.0.1
django-simple-email-confirmation==0.60
django-taggit==1.1.0
django-tinymce==2.8.0
django-verified-email-field==1.5.1
django==2.2.8
eth-abi==2.0.0
eth-account==0.4.0
eth-hash[pycryptodome]==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.1.2
eth-typing==2.2.1
eth-utils==1.8.1
geographiclib==1.50
geopy==1.20.0
grpcio-tools==1.25.0
grpcio==1.25.0
hexbytes==0.2.0
idna==2.8
importlib-metadata==1.1.0 ; python_version < '3.8'
ipfshttpclient==0.4.12
isort==4.3.21
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lru-dict==1.1.6
luhn==0.2.0
mccabe==0.6.1
mnemonic==0.19
more-itertools==8.0.0
multiaddr==0.0.8
mypy-extensions==0.4.3
netaddr==0.7.19
parsimonious==0.8.1
phonenumbers==8.11.0
pillow==6.2.1
protobuf==3.11.1
psycopg2==2.8.4
pycryptodome==3.9.4
pyjwt==1.7.1
pylint==2.4.4
pyrsistent==0.15.6
pytz==2019.3
redis==3.3.11
requests==2.22.0
rlp==1.2.0
six==1.13.0
sqlparse==0.3.0
stop-words==2018.7.23
toolz==0.10.0
twilio==6.33.1
typed-ast==1.4.0 ; implementation_name == 'cpython' and python_version < '3.8'
urllib3==1.25.7
varint==1.0.2
web3==5.3.0
websockets==8.1
wrapt==1.11.2
zipp==0.6.0
